requirements.txt
aiogram>=3.24.0
aiosqlite>=0.21.0
httpx>=0.28.0
apscheduler>=3.11.0
python-dotenv>=1.1.0
pydantic>=2.11.0
